refactor: sealed class를 이용한 viewHolder 로직 수정
